Rename component, change prop type
