+++ src/views/Accessibility/AccessibiltyRequest/TestDate/index.stories.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { DateTime } from 'luxon';
+
+import TestDate from './index';
+
+const Template = (args: any) => (
+  <div className="grid-col-4">
+    <TestDate {...args} />
+  </div>
+);
+
+export default {
+  title: 'Test Date',
+  component: TestDate,
+  parameters: {
+    info: 'The Test Date component'
+  },
+  args: {
+    date: DateTime.local()
+  }
+};
+
+export const Default = Template.bind({});
