+++ src/views/Accessibility/AccessibiltyRequest/TestDate/index.tsx
@@ -1,21 +1,24 @@
 import React from 'react';
 import { Link } from '@trussworks/react-uswds';
+import { DateTime } from 'luxon';
 
-type TestDatesProps = {
-  date: string;
+type TestDateProps = {
+  date: DateTime;
   isInitial: boolean;
   testIndex: number;
   score?: string; // This type might need to be changed based on API model
 };
 
-const TestDates = ({ date, isInitial, testIndex, score }: TestDatesProps) => {
+const TestDate = ({ date, isInitial, testIndex, score }: TestDateProps) => {
   return (
     <div className="bg-gray-10 padding-2 line-height-body-4">
       <div className="text-bold margin-bottom-1">
         Test {testIndex}: {isInitial ? 'Initial' : 'Remediation'}
       </div>
       <div className="margin-bottom-1">
-        <div className="display-inline-block margin-right-2">{date}</div>
+        <div className="display-inline-block margin-right-2">
+          {date.toLocaleString(DateTime.DATE_FULL)}
+        </div>
         <div
           className="display-inline-block text-base-dark"
           data-testid="score"
@@ -33,4 +36,4 @@ const TestDates = ({ date, isInitial, testIndex, score }: TestDatesProps) => {
   );
 };
 
-export default TestDates;
+export default TestDate;
